Refactor settings_page_list_tile_two.dart and SettingsPageListTileTwo.
