+++ lib/development/bu8-settings_page_contents.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 //  Import project-specific files.
 import 'package:kar_kam/lib/data_notifier.dart';
-import 'package:kar_kam/settings_page_list_tile.dart';
+import 'package:kar_kam/development/bu10-settings_page_list_tile.dart';
 
 
 class SettingsPageContents extends StatefulWidget {
@@ -60,7 +60,7 @@ class _SettingsPageContentsState extends State<SettingsPageContents> {
             );
           }),
           ...List<Widget>.generate(1, (int index) {
-            return SettingsPageListTile();
+            return AlternativeSettingsPageListTile();
           }),
           ...List<Widget>.generate(15, (int index) {
             return Opacity(
