+++ lib/settings_page_contents.dart
@@ -2,8 +2,8 @@
 import 'package:flutter/material.dart';
 
 //  Import project-specific files.
+import 'package:kar_kam/development/bu10-settings_page_list_tile.dart';
 import 'package:kar_kam/settings_page_list_tile.dart';
-import 'package:kar_kam/settings_page_list_tile_two.dart';
 import 'package:kar_kam/lib/data_notifier.dart';
 
 //  [SettingsPageContents] creates app settings in the form of a ListView.
@@ -67,11 +67,8 @@ class _SettingsPageContentsState extends State<SettingsPageContents> {
               ),
             );
           }),
-          ...List<Widget>.generate(1, (int index) {
-            return SettingsPageListTile();
-          }),
           ...List<Widget>.generate(100, (int index) {
-            return SettingsPageListTileTwo();
+            return SettingsPageListTile();
           }),
           ...List<Widget>.generate(15, (int index) {
             return Opacity(
