+++ lib/settings_page_list_tile.dart
@@ -2,39 +2,50 @@
 import 'package:flutter/material.dart';
 
 // Import project-specific files.
-import 'package:kar_kam/clipped_rounded_rectangle_border.dart';
 import 'package:kar_kam/lib/data_notifier.dart';
+import 'package:kar_kam/settings_page_list_tile_clipper.dart';
 
-class SettingsPageListTile extends StatelessWidget {
+/// [SettingsPageListTile] implements a draft listTile effect that is
+/// able to morph itself and pass around ButtonArray.
+class SettingsPageListTile extends StatefulWidget {
   const SettingsPageListTile({Key? key}) : super(key: key);
 
   @override
+  State<SettingsPageListTile> createState() => _SettingsPageListTileState();
+}
+
+class _SettingsPageListTileState extends State<SettingsPageListTile> {
+  @override
   Widget build(BuildContext context) {
-    //  Retrieve buttonArrayRect from instance of DataNotifier
+    //  Retrieve [buttonArrayRect] from instance of DataNotifier
     //  further up the widget tree.
+    //
+    //  [buttonArrayRect] is passed to [SettingsPageListTileClipper] which
+    //  calculates the path for ClipPath to use.
     Rect buttonArrayRect = DataNotifier
         .of(context, ValueKey('buttonArrayRect')).data.value;
-    
+
+    //  Construct a morphable ListTile instance using ClipPath.
+    //
+    //  ValueListenableBuilder rebuilds whenever the scroll position changes,
+    //  but the build doesn't depend on it.
     return ValueListenableBuilder<double>(
       valueListenable: DataNotifier
           .of(context, ValueKey('scrollController')).data,
       builder: (BuildContext context, double value, __) {
-        // print('SettingsPageListTile, value = $value');
-        return Card(
-          shape: ClippedRoundedRectangleBorder(
-            pos: value,
+        return ClipPath(
+          clipper: SettingsPageListTileClipper(
             context: context,
-            side: BorderSide(width: 2.0, color: Colors.black, style: BorderStyle.solid),
-            // borderRadius: BorderRadius.circular(10+ value),
-            borderRadius: BorderRadius.circular(10),
             guestRect: buttonArrayRect,
           ),
-          elevation: 20,
-          color: Colors.amber,
-          child: const ListTile(
-            leading: FlutterLogo(size: 72.0),
-            title: Text('SettingsPageListTile'),
-            trailing: Icon(Icons.more_vert),
+          child: Card(
+            elevation: 20,
+            color: Colors.amber,
+            child: const ListTile(
+              leading: FlutterLogo(size: 72.0),
+              title: Text('SettingsPageListTile'),
+              trailing: Icon(Icons.more_vert),
+            ),
           ),
         );
       },
